Parallelise all the things.
Requests to the API are now created in advance, then executed in parallel by grequests.

The context middleware no longer requests /whoami; the request for this is instead constructed
and added to request.view_requests, a list. In view functions, any further API requests should
be appended to this list, then the entire list of requests can be executed with grequests.map(request.view_requests).

If there is only one request made to the API for a page load, this actually adds overhead.
But in the majority of page loads more than one request is made, so it is advantageous.