Add 'move' moderation action for items
This allows the moderator of a microcosm to move single items to a
different microcosm.

Display of the moderation menu is based on the 'moderator' permissions
flag.

Some edge cases (such as moving an item to a microcosm which is private)
are not yet handled.

Styling of the forms is still to be done, but they function.